190423 [add]fakeApi_kari
